requirements.txt
django==3.1
djangorestframework==3.12.1
markdown==3.3.1
django-filter==2.4.0
django-softdelete==0.9.2
django-celery-beat==1.5.0
celery==4.3.0
vine==1.3.0
psycopg2-binary==2.7.6.1
redis==3.5.3
scipy
sklearn
biopython==1.60

# development
ipython

